refactor: useEffect 간소화, 컴포넌트 구조 및 오타 수정
